New SearchText test focusing on the stack overflow question
PDF Clown Highlight multiple search word is failing for PDF contains images, color text, Complex Diagrams
https://stackoverflow.com/questions/45498469/pdf-clown-highlight-multiple-search-word-is-failing-for-pdf-contains-images-col
751-07-T-A-2X-1-101.pdf
https://drive.google.com/file/d/0B-nuOO6Zsa4rXy1DS2JjX1RnYmM/view?usp=sharing

As it turns out, the search fails due to a bug in PDF Clown:

For sorting it uses a custom Comparator implementation, the TextExtractor.TextStringPositionComparator. Unfortunately this implementation does not fulfill the Comparator contract. Former Java versions ignored this but the TimSort by default used in Java 8 recognizes this and fails.

If this test is run with the "-Djava.util.Arrays.useLegacyMergeSort=true", the legacy sorting method is used instead and the test succeeds.