New ExtractSize test focusing on the stack overflow question
PDFClown Different font-size in one line
https://stackoverflow.com/questions/45413682/pdfclown-different-font-size-in-one-line
SameCharactersDifferentSize.pdf
https://www.file-upload.net/download-12650542/SameCharactersDifferentSize.pdf.html

The issue is caused by a bug in PDF Clown: it assumes that marked content sections and save/restore graphics state blocks are properly contained in each other and don't overlap.

The only easy way I found to repair this was to disable the marked content parsing in PDF Clown.