New HighlightChinese test focusing on the stack overflow question
pdf clown- not highlighting specific search keyword
https://stackoverflow.com/questions/48395771/pdf-clown-not-highlighting-specific-search-keyword
chinese1.pdf - via e-mail

Indeed, there are issues with the example PDF but they are different from the OP's description. This might be due to him not using the current PDF Clown version. The issues observed are due to an error in the PDF Clown ShowText class. A special treatment of ShowTextToNextLine instances at the end of its scan method results in errors calculating the text line matrix and as a result later text matrices.